% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{agridata_options}
\alias{agridata_options}
\title{Set or get options for ragridata}
\usage{
agridata_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\value{
the default and modified options.
}
\description{
Set or get options for ragridata
}
\section{Supported options}{

The following options are supported
}

\examples{
agridata_options()
}
