% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{agridata_reset}
\alias{agridata_reset}
\title{Reset global options for ragridata}
\usage{
agridata_reset()
}
\value{
the default options
}
\description{
Reset global options for ragridata
}
\examples{
agridata_reset()
}
