% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asris.R
\name{asris_soil}
\alias{asris_soil}
\title{Download soil file from asris and convert to json which can be used in APSIM NG}
\usage{
asris_soil(lon, lat, outfile, models)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{outfile}{output file}

\item{models}{path to models}
}
\value{
No return values
}
\description{
Download soil file from asris and convert to json which can be used in APSIM NG
}
