% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo.R
\name{silo_grid}
\alias{silo_grid}
\title{Get gridded data from SILO}
\usage{
silo_grid(lon, lat, startDate, endDate, email, outfile)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{startDate}{start date}

\item{endDate}{end date}

\item{email}{email address}

\item{outfile}{output file}
}
\value{
No return values
}
\description{
Get gridded data from SILO
}
\details{
Get weather record from SILO Gridded Dataset (https://www.longpaddock.qld.gov.au/silo/api-documentation/)
}
